+++ packages/genesys/packages/react-components-core/src/Alert/Alert.types.ts
@@ -1,15 +1,15 @@
-import { ReactElement } from "react";
+import { ReactElement, ReactNode } from "react";
 import { CorePaperProps } from "../Paper";
 
 export type AlertType = "info" | "success" | "error" | "warning" | "loading";
 
 export interface CoreAlertProps extends Omit<CorePaperProps, "children"> {
     /**
-     * Alert message
+     * Alert content
      */
-    message: string;
+    content: ReactNode;
     /**
-     * Alert message
+     * Alert type (status: "info" | "success" | "error" | "warning" | "loading")
      */
     type?: AlertType;
     /**